requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic-settings==2.0.3
sqlalchemy==1.4.48
PyMySQL==1.1.0
pymongo==4.5.0
motor==3.3.1
python-dotenv==1.0.0
